Refactor responses to use a common struct
